% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bar}
\alias{bar}
\title{Simple example function}
\usage{
bar(s = "baz")
}
\arguments{
\item{s}{A character string, defaults to \dQuote{baz}}
}
\value{
Nothing is returned as the function is invoked for its side-effect.
}
\description{
The \pkg{tl} package provides a uniform interface to logging
from both R and C++ using the same namespaced \code{tl::debug()} (and alike
for differ logging verbs) interface. This package (and function)
demonstrates it.
}
