% tinyrox says don't edit this manually, but it can't stop you!
\name{foo}
\alias{foo}
\title{Simple Example of tl() use}
\usage{
foo()
}
\description{
This function details how one can combine logging from R (by
calling functions such as \code{tl::debug()}) with C++ logging, here from a
function \code{bar()} in this package.
}
\examples{
foo()
}
