\name{tlexample-package}
\alias{tlexample-package}
\alias{tlexample}
\docType{package}
\title{\packageTitle{tlexample}}
\description{\packageDescription{tlexample}}
\section{Package Content}{\packageIndices{tlexample}}
\author{\packageAuthor{tlexample}}
\section{Maintainer}{\packageMaintainer{tlexample}}
\keyword{package}
